feat(signals,solid,universal): ONE list engine — mapArray is the engine's array output; solid-js keeps the node layer
The row core moves into @solidjs/signals (list.ts). `mapArray` is now a
thin wrapper over its ARRAY output (same API and contract: values, array
identity while structurally unchanged, [fallback] when empty, dev strict-
read name, `_parentComputed` routing so untracked store reads in row bodies
see pending writes). `<For>`'s plain call returns the same. The old mapArray
body is deleted from the shipped package and frozen as a TEST REFERENCE
(web/test/reference/mapArray.ts, on public API) — the oracle every engine
harness now compares against.

The RENDERED output drives the engine through a ListNodeLayer — the only
code that touches nodes. solid-js (client/for-slot.ts) builds that layer
over a renderer's SlotOps (value→nodes, dynamic-row resolve/same/splice,
place/detach/anchors/bulk clear/hydration commit) and wires the render
effect. Web (domOps) and universal (createRenderer primitives) both
consume the same layer — universal's oracle test extended with dynamic
rows as the second-renderer proof. When the layer is null (array output)
nothing node-related is referenced.

Two bugs the reference oracle caught in the move itself, both fixed:
- listArray created its computed under the READER (a <For> builds array
  output lazily on first read) → disposed with the reader's next run.
  Created under the list's creation owner now.
- Routing row owners' _parentComputed through the RENDERED compute gave
  row-created memos (dynamic rows) a height above the compute → height
  inversion, double compute per flush on a dynamic flip + list change,
  rebuilding the pass's fresh rows. Routing is array-output only.

Suites: web 810, server 779, hydrate 187, solid 585, universal 47,
signals 1503 (its 35 mapArray tests now run against the engine).
Chrome (interleaved, pre-move vs moved): reorder and creates identical.

Size (brotli, vs next): For-only app +3.81 KB (layer boundary ~+220 B);
For + mapArray app +3.87 KB (was +4.38: no second list implementation);
mapArray without For +3.69 KB (was +0.85 — the rendered-only control flow
shares the engine closure; pluggable flat/LIS/dynamic/placement is the
follow-up if that consumer shape matters).
